+++ packages/insomnia/src/ui/routes/actions.tsx
@@ -5,39 +5,25 @@ import { generate, runTests, type Test } from 'insomnia-testing';
 import type { TestResults } from 'insomnia-testing/src/run/entities';
 import { type ActionFunction, redirect } from 'react-router';
 
-import { version } from '../../../package.json';
 import { parseApiSpec, resolveComponentSchemaRefs } from '../../common/api-specs';
-import { ACTIVITY_DEBUG, getAIServiceURL, METHOD_GET } from '../../common/constants';
+import { ACTIVITY_DEBUG, getAIServiceURL } from '../../common/constants';
 import { database } from '../../common/database';
-import {
-  importResourcesToNewWorkspace,
-  importResourcesToWorkspace,
-  scanResources,
-  type ScanResult,
-} from '../../common/import';
-import { getInsomniaV5DataExport, importInsomniaV5Data } from '../../common/insomnia-v5';
+import { importResourcesToWorkspace, scanResources, type ScanResult } from '../../common/import';
 import { generateId } from '../../common/misc';
 import * as models from '../../models';
 import { EnvironmentType } from '../../models/environment';
 import { getById, update } from '../../models/helpers/request-operations';
-import type { MockServer } from '../../models/mock-server';
-import { isGitProject, isRemoteProject, type Project } from '../../models/project';
+import { isGitProject } from '../../models/project';
 import { isRequest, type Request } from '../../models/request';
 import { isRequestGroup, isRequestGroupId } from '../../models/request-group';
 import { isRequestGroupMeta } from '../../models/request-group-meta';
 import type { UnitTest } from '../../models/unit-test';
 import type { UnitTestSuite } from '../../models/unit-test-suite';
-import { isCollection, isEnvironment, scopeToActivity, type Workspace } from '../../models/workspace';
 import type { WorkspaceMeta } from '../../models/workspace-meta';
 import { getSendRequestCallback } from '../../network/unit-test-feature';
-import { initializeLocalBackendProjectAndMarkForSync } from '../../sync/vcs/initialize-backend-project';
-import { VCSInstance } from '../../sync/vcs/insomnia-sync';
 import { insomniaFetch } from '../../ui/insomniaFetch';
 import { invariant } from '../../utils/invariant';
 import { SegmentEvent } from '../analytics';
-import { syncNewWorkspaceIfNeeded } from './import';
-
-// Project
 
 export function safeToUseInsomniaFileName(fileName: string) {
   const fileNameWithoutExt = fileName.replace('.yaml', '').replace('.yml', '');
@@ -53,327 +39,6 @@ export function safeToUseInsomniaFileName(fileName: string) {
 export function safeToUseInsomniaFileNameWithExt(fileName: string) {
   return `${safeToUseInsomniaFileName(fileName)}.yaml`;
 }
-
-// Workspace
-export const createNewWorkspaceAction: ActionFunction = async ({ params, request }) => {
-  const { organizationId, projectId } = params;
-  invariant(organizationId, 'Organization ID is required');
-  invariant(projectId, 'Project ID is required');
-
-  const project = await models.project.getById(projectId);
-
-  invariant(project, 'Project not found');
-
-  const formData = await request.formData();
-
-  const name = formData.get('name');
-
-  invariant(typeof name === 'string', 'Name is required');
-
-  const scope = formData.get('scope');
-  invariant(
-    scope === 'design' || scope === 'collection' || scope === 'mock-server' || scope === 'environment',
-    'Scope is required',
-  );
-
-  const flushId = await database.bufferChanges();
-
-  const workspaceName = name || (scope === 'collection' ? 'My Collection' : 'my-spec.yaml');
-
-  const workspace = await models.workspace.create({
-    name: workspaceName,
-    scope,
-    parentId: projectId,
-  });
-
-  if (isGitProject(project)) {
-    const workspaceMeta = await models.workspaceMeta.getOrCreateByParentId(workspace._id);
-
-    const fileName = formData.get('fileName')?.toString() || workspace.name;
-
-    const safeToUseFileNameWithExtension = safeToUseInsomniaFileNameWithExt(fileName);
-
-    await models.workspaceMeta.update(workspaceMeta, {
-      gitFilePath: path.join(formData.get('folderPath')?.toString() || '', safeToUseFileNameWithExtension),
-    });
-  }
-
-  if (scope === 'mock-server') {
-    const mockServerType = formData.get('mockServerType');
-    invariant(mockServerType === 'cloud' || mockServerType === 'self-hosted', 'Mock Server type is required');
-
-    const mockServerPatch: Partial<MockServer> = {
-      name,
-    };
-
-    if (mockServerType === 'cloud') {
-      mockServerPatch.useInsomniaCloud = true;
-    }
-
-    if (mockServerType === 'self-hosted') {
-      const mockServerUrl = formData.get('mockServerUrl');
-      invariant(typeof mockServerUrl === 'string', 'Mock Server URL is required');
-      mockServerPatch.useInsomniaCloud = false;
-      mockServerPatch.url = mockServerUrl;
-    }
-
-    await models.environment.getOrCreateForParentId(workspace._id);
-    const workspaceMeta = await models.workspaceMeta.getOrCreateByParentId(workspace._id);
-    await models.mockServer.getOrCreateForParentId(workspace._id, mockServerPatch);
-    await database.flushChanges(flushId);
-
-    const { id } = await models.userSession.getOrCreate();
-    if (id && !workspaceMeta.gitRepositoryId) {
-      const vcs = VCSInstance();
-      await initializeLocalBackendProjectAndMarkForSync({
-        vcs,
-        workspace,
-      });
-    }
-    window.main.trackSegmentEvent({
-      event: SegmentEvent.mockCreate,
-    });
-    return redirect(
-      `/organization/${organizationId}/project/${projectId}/workspace/${workspace._id}/${scopeToActivity(workspace.scope)}`,
-    );
-  }
-
-  if (scope === 'design') {
-    await models.apiSpec.getOrCreateForParentId(workspace._id);
-  }
-
-  // Create default env, cookie jar, and meta
-  await models.environment.getOrCreateForParentId(workspace._id);
-  await models.cookieJar.getOrCreateForParentId(workspace._id);
-  const workspaceMeta = await models.workspaceMeta.getOrCreateByParentId(workspace._id);
-
-  await database.flushChanges(flushId);
-
-  const { id } = await models.userSession.getOrCreate();
-  if (id && !workspaceMeta.gitRepositoryId && !isGitProject(project)) {
-    const vcs = VCSInstance();
-    await initializeLocalBackendProjectAndMarkForSync({
-      vcs,
-      workspace,
-    });
-  }
-
-  let event = SegmentEvent.documentCreate;
-
-  if (isCollection(workspace)) {
-    event = SegmentEvent.collectionCreate;
-  } else if (isEnvironment(workspace)) {
-    event = SegmentEvent.environmentWorkspaceCreate;
-  }
-
-  window.main.trackSegmentEvent({
-    event: event,
-  });
-
-  // Parse the URL from the request
-  const url = new URL(request.url);
-
-  // Get a specific query parameter
-  const withRequest = url.searchParams.get('withRequest');
-
-  if (withRequest) {
-    const settings = await models.settings.getOrCreate();
-    const defaultHeaders = settings.disableAppVersionUserAgent
-      ? []
-      : [{ name: 'User-Agent', value: `insomnia/${version}` }];
-
-    const activeRequestId = (
-      await models.request.create({
-        parentId: workspace._id,
-        method: METHOD_GET,
-        name: 'My first request',
-        headers: defaultHeaders,
-      })
-    )._id;
-
-    window.main.trackSegmentEvent({ event: SegmentEvent.requestCreate, properties: { requestType: 'HTTP' } });
-
-    return redirect(
-      `/organization/${organizationId}/project/${projectId}/workspace/${workspace._id}/debug/request/${activeRequestId}`,
-    );
-  }
-
-  return redirect(
-    `/organization/${organizationId}/project/${projectId}/workspace/${workspace._id}/${scopeToActivity(workspace.scope)}`,
-  );
-};
-
-async function deleteWorkspaceFromCloud(workspace: Workspace, project: Project) {
-  const workspaceMeta = await models.workspaceMeta.getOrCreateByParentId(workspace._id);
-  const isGitSync = !!workspaceMeta.gitRepositoryId;
-
-  if (isRemoteProject(project) && !isGitSync) {
-    try {
-      const vcs = VCSInstance();
-      await vcs.switchAndCreateBackendProjectIfNotExist(workspace._id, workspace.name);
-      await vcs.archiveProject();
-    } catch (err) {
-      return {
-        error:
-          err instanceof Error
-            ? err.message
-            : `An unexpected error occurred while deleting the workspace. Please try again. ${err}`,
-      };
-    }
-  }
-
-  return null;
-}
-
-async function deleteWorkspaceFromLocal(workspace: Workspace) {
-  await models.stats.incrementDeletedRequestsForDescendents(workspace);
-  await models.workspace.remove(workspace);
-}
-
-async function deleteWorkspace(workspace: Workspace | null, project: Project | null) {
-  invariant(workspace, 'Workspace not found');
-  invariant(project, 'Project not found');
-
-  const ret = await deleteWorkspaceFromCloud(workspace, project);
-  if (ret?.error) {
-    return ret;
-  }
-
-  await deleteWorkspaceFromLocal(workspace);
-
-  return null;
-}
-
-export const deleteWorkspaceAction: ActionFunction = async ({ params, request }) => {
-  const { organizationId, projectId } = params;
-  invariant(projectId, 'projectId is required');
-
-  const project = await models.project.getById(projectId);
-
-  const formData = await request.formData();
-
-  const workspaceId = formData.get('workspaceId');
-  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
-
-  const workspace = await models.workspace.getById(workspaceId);
-
-  const msgObj = await deleteWorkspace(workspace, project);
-
-  if (msgObj?.error) {
-    return msgObj;
-  }
-
-  return redirect(`/organization/${organizationId}/project/${projectId}`);
-};
-
-/** Duplicate workspace to other project and automatically sync to cloud if needed  */
-export const duplicateWorkspaceAction: ActionFunction = async ({ request }) => {
-  const formData = await request.formData();
-  const oldWorkspaceId = formData.get('workspaceId') as string;
-  invariant(oldWorkspaceId, 'Workspace ID is required');
-  const newOrgId = formData.get('orgId') as string;
-  invariant(newOrgId, 'Org ID is required');
-  const newProjectId = formData.get('projectId') as string;
-  invariant(newProjectId, 'Project ID is required');
-  const newWorkspaceName = formData.get('name') as string;
-
-  const oldWorkspace = await models.workspace.getById(oldWorkspaceId);
-  invariant(oldWorkspace, 'Workspace not found');
-
-  // duplicate the workspace to the new project
-  const newProject = (await models.project.getById(newProjectId)) as Project;
-  const workspaceExport = await getInsomniaV5DataExport({
-    workspaceId: oldWorkspace._id,
-    includePrivateEnvironments: true,
-  });
-
-  const data = importInsomniaV5Data(workspaceExport);
-
-  const newWorkspace = await importResourcesToNewWorkspace({
-    projectId: newProject._id,
-    workspaceToImport: {
-      ...oldWorkspace,
-      name: newWorkspaceName || oldWorkspace.name,
-    },
-    resourceCacheItem: {
-      resources: data,
-      content: JSON.stringify(data, null, 2),
-      importer: {
-        id: 'insomnia-v5',
-        name: 'Insomnia v5 Importer',
-        description: 'Import Insomnia v5 data',
-      },
-    },
-    syncNewWorkspaceIfNeeded,
-  });
-  return redirect(
-    `/organization/${newOrgId}/project/${newProjectId}/workspace/${newWorkspace._id}/${scopeToActivity(newWorkspace.scope)}`,
-  );
-};
-
-export const updateWorkspaceAction: ActionFunction = async ({ request }) => {
-  const patch = await request.json();
-  const workspaceId = patch.workspaceId;
-  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
-  const workspace = await models.workspace.getById(workspaceId);
-  invariant(workspace, 'Workspace not found');
-
-  if (workspace.scope === 'design') {
-    const apiSpec = await models.apiSpec.getByParentId(workspaceId);
-    invariant(apiSpec, 'No Api Spec found for this workspace');
-
-    await models.apiSpec.update(apiSpec, {
-      fileName: patch.name || workspace.name,
-    });
-  }
-
-  if (workspace.scope === 'mock-server') {
-    const mockServer = await models.mockServer.getByParentId(workspaceId);
-    invariant(mockServer, 'No MockServer found for this workspace');
-
-    let useInsomniaCloud = mockServer.useInsomniaCloud;
-    if (patch.mockServerType && typeof patch.mockServerType === 'string') {
-      useInsomniaCloud = patch.mockServerType === 'cloud';
-    }
-
-    let mockServerUrl = mockServer.url;
-
-    if (patch.mockServerUrl && typeof patch.mockServerUrl === 'string') {
-      mockServerUrl = patch.mockServerUrl;
-    }
-
-    await models.mockServer.update(mockServer, {
-      name: patch.name || workspace.name,
-      useInsomniaCloud,
-      url: mockServerUrl,
-    });
-  }
-
-  patch.name = patch.name || workspace.name || (workspace.scope === 'collection' ? 'My Collection' : 'my-spec.yaml');
-
-  await models.workspace.update(workspace, patch);
-
-  const project = await models.project.getById(workspace.parentId);
-  invariant(project, 'Project not found');
-  if (isGitProject(project)) {
-    const workspaceMeta = await models.workspaceMeta.getOrCreateByParentId(workspace._id);
-
-    const existingPathDir = path.dirname(workspaceMeta.gitFilePath || '');
-    let fileName = path.basename(workspaceMeta.gitFilePath || '');
-
-    if (patch.fileName && typeof patch.fileName === 'string') {
-      fileName = patch.fileName;
-    }
-
-    await models.workspaceMeta.update(workspaceMeta, {
-      gitFilePath: path.join(existingPathDir, safeToUseInsomniaFileNameWithExt(fileName)),
-    });
-  }
-
-  return {
-    success: true,
-  };
-};
 
 export const moveWorkspaceIntoProjectAction: ActionFunction = async ({ request, params }) => {
   const { organizationId } = params;
